requirements.txt
asgiref==3.3.4
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
decorator==4.4.2
Django==3.2
django-formtools==2.3
django-otp==1.0.3
django-otp-yubikey==1.0.0.post1
django-phonenumber-field==5.1.0
django-two-factor-auth==1.13.1
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
dnspython==1.16.0
docopt==0.6.2
docutils==0.17.1
fett==0.3.2
idna==2.10
networkx==2.5.1
phonenumbers==8.12.21
pycryptodome==3.10.1
PyJWT==1.7.1
pymongo==3.11.3
python-jsonrpc-server==0.3.4
pytz==2021.1
PyYAML==5.4.1
qrcode==6.1
requests==2.25.1
six==1.15.0
snooty-lextudio==1.9.0.dev0
sqlparse==0.4.1
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.26.4
watchdog==1.0.2
YubiOTP==1.0.0
